requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-js-asset==1.2.2
django-recaptcha==2.0.6
django-taggit==1.3.0
mysqlclient==2.0.2
Pillow==8.0.1
PyMySQL==0.10.1
pytz==2020.5
sqlparse==0.4.1
